requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.7.0
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
discord.py==1.7.1
idna==2.10
kaleido==0.2.1
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.1
multidict==5.1.0
multitasking==0.0.9
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
six==1.15.0
toml==0.10.2
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
yarl==1.6.3
yfinance==0.1.59
